release staging

+++ app/Http/Controllers/Api/PatientSouscripteurController.php
@@ -64,9 +64,12 @@ class PatientSouscripteurController extends Controller
         }
 
         $patient = Patient::with(['souscripteur.user','user','affiliations','etablissements','financeurs.financable.user'])->restrictUser()->whereSlug($patient->slug)->first();
-        if (!is_null($patient->user->email)){
-            $mail = new updateSetting($patient->user);
-            Mail::to($patient->user->email)->send($mail);
+        $user = $patient->user;
+        if ($user->decede == 'non') {
+            if (!is_null($patient->user->email)) {
+                $mail = new updateSetting($patient->user);
+                Mail::to($patient->user->email)->send($mail);
+            }
         }
         //Notification de mises à jour de compte
 
@@ -83,56 +86,58 @@ class PatientSouscripteurController extends Controller
     public function retirer(Request $request){
 
         $patient = Patient::find($request->get('patient_id'));
-        if (!empty($request->get('souscripteurs'))){
-            $souscripteur = Souscripteur::whereSlug($request->get('souscripteurs')[0])->first();
-            if ($patient->souscripteur_id == $souscripteur->user_id){
-                $patient->souscripteur_id = null;
-                $patient->save();
+        $user = $patient->user;
+        if ($user->decede == 'non') {
+            if (!empty($request->get('souscripteurs'))) {
+                $souscripteur = Souscripteur::whereSlug($request->get('souscripteurs')[0])->first();
+                if ($patient->souscripteur_id == $souscripteur->user_id) {
+                    $patient->souscripteur_id = null;
+                    $patient->save();
 
-                if (!is_null($patient->souscripteur)){
-                    if (!is_null($patient->souscripteur->user)){
-                        $mail = new updateSetting($patient->souscripteur->user);
-                        Mail::to($patient->souscripteur->user->email)->send($mail);
+                    if (!is_null($patient->souscripteur)) {
+                        if (!is_null($patient->souscripteur->user)) {
+                            $mail = new updateSetting($patient->souscripteur->user);
+                            Mail::to($patient->souscripteur->user->email)->send($mail);
+                        }
                     }
                 }
-            }
 
 //            defineAsAuthor('PatientSouscripteur',$financeur->id,'create',$patient->user_id);
 
-        }
-        if (!empty($request->get('financable_slug'))){
-            $financeurSlug = $request->get('financable_slug');
-            foreach ($financeurSlug  as $slug){
-                $this->validatedSlug($slug,$this->table);
-                $financeur = PatientSouscripteur::whereSlug($slug)->first();
-                $financeur->delete();
+            }
+            if (!empty($request->get('financable_slug'))) {
+                $financeurSlug = $request->get('financable_slug');
+                foreach ($financeurSlug as $slug) {
+                    $this->validatedSlug($slug, $this->table);
+                    $financeur = PatientSouscripteur::whereSlug($slug)->first();
+                    $financeur->delete();
 
-                defineAsAuthor('PatientSouscripteur',$financeur->id,'create',$patient->user_id);
+                    defineAsAuthor('PatientSouscripteur', $financeur->id, 'create', $patient->user_id);
 
-                if (!is_null($financeur->financable)){
-                    if (!is_null($financeur->financable->user)){
-                        $mail = new updateSetting($financeur->financable->user);
-                        Mail::to($financeur->financable->user->email)->send($mail);
+                    if (!is_null($financeur->financable)) {
+                        if (!is_null($financeur->financable->user)) {
+                            $mail = new updateSetting($financeur->financable->user);
+                            Mail::to($financeur->financable->user->email)->send($mail);
+                        }
                     }
                 }
             }
-        }
 
-        $patient = Patient::with(['souscripteur.user','user','affiliations','etablissements','financeurs.financable.user'])->restrictUser()->whereSlug($patient->slug)->first();
+            $patient = Patient::with(['souscripteur.user', 'user', 'affiliations', 'etablissements', 'financeurs.financable.user'])->restrictUser()->whereSlug($patient->slug)->first();
 
-        //Notification de mises à jour de compte
-        try{
-            if (!is_null($patient->user->email)){
-                $mail = new updateSetting($patient->user);
-                Mail::to($patient->user->email)->send($mail);
+            //Notification de mises à jour de compte
+            try {
+                if (!is_null($patient->user->email)) {
+                    $mail = new updateSetting($patient->user);
+                    Mail::to($patient->user->email)->send($mail);
+                }
+
+            } catch (\Swift_TransportException $transportException) {
+                $message = "L'operation à reussi mais le mail n'a pas ete envoye. Verifier votre connexion internet ou contacter l'administrateur";
+                return response()->json(['patient' => $patient, "message" => $message]);
+
             }
-
-        }catch (\Swift_TransportException $transportException){
-            $message = "L'operation à reussi mais le mail n'a pas ete envoye. Verifier votre connexion internet ou contacter l'administrateur";
-            return response()->json(['patient'=>$patient, "message"=>$message]);
-
         }
-
         return response()->json(['patient'=>$patient]);
     }
